+++ packages/nhsuk-frontend/src/nhsuk/components/action-link/fixtures.mjs
@@ -29,7 +29,7 @@ const fixtures = {
   "default": {
     context: {
       text: "Find your nearest A&E",
-      href: "#"
+      href: "#/find"
     },
     variants,
     screenshot: {
@@ -52,14 +52,19 @@ const fixtures = {
     context: {
       html: outdent`
         Start session<br>
-        <span class="nhsuk-u-font-weight-normal nhsuk-u-font-size-19">(11 cases)</span>
-      `
-    }
+        <span class="nhsuk-u-secondary-text-colour nhsuk-u-font-weight-normal nhsuk-u-font-size-19">(11 cases)</span>
+      `,
+      href: "#/start"
+    },
+    variants
   },
   "with HTML via call block": {
+    context: {
+      href: "#/start"
+    },
     callBlock: outdent`
       Start session<br>
-      <span class="nhsuk-u-font-weight-normal nhsuk-u-font-size-19">(11 cases)</span>
+      <span class="nhsuk-u-secondary-text-colour nhsuk-u-font-weight-normal nhsuk-u-font-size-19">(11 cases)</span>
     `
   }
 }